refactor: moved backoff_heap
